+++ src/core/log.cpp
@@ -5,51 +5,35 @@
 // Copyright: (c) 2020 Anton van Wezenbeek
 ////////////////////////////////////////////////////////////////////////////////
 
-#include <easylogging++.h>
+#include <boost/log/core.hpp>
+#include <boost/log/expressions.hpp>
+#include <boost/log/trivial.hpp>
+#include <boost/log/utility/setup/common_attributes.hpp>
+#include <boost/log/utility/setup/console.hpp>
+#include <boost/log/utility/setup/file.hpp>
+
 #include <iomanip>
+#include <wex/cmdline.h>
 #include <wex/config.h>
 #include <wex/log.h>
 #include <wex/path.h>
+#include <wx/app.h>
 #include <wx/log.h>
 
-INITIALIZE_EASYLOGGINGPP
+namespace logging = boost::log;
 
 enum level_t
 {
-  // boost
-  LEVEL_FATAL = 1,
-  LEVEL_WARNING,
-  LEVEL_INFO,
-  LEVEL_DEBUG,
+  // follow boost
   LEVEL_TRACE,
-
+  LEVEL_DEBUG,
+  LEVEL_INFO,
+  LEVEL_WARNING,
   LEVEL_ERROR,
-  // other levels
+  LEVEL_FATAL,
+  // wxLog
   LEVEL_STATUS,
 };
-
-const std::string level_string(int level)
-{
-  switch (level)
-  {
-    case LEVEL_DEBUG:
-      return "debug";
-    case LEVEL_ERROR:
-      return "error";
-    case LEVEL_FATAL:
-      return "fatal";
-    case LEVEL_INFO:
-      return "info";
-    case LEVEL_STATUS:
-      return "status";
-    case LEVEL_TRACE:
-      return "trace";
-    case LEVEL_WARNING:
-      return "warning";
-    default:
-      return std::to_string(level);
-  }
-}
 
 wex::log::log(const std::string& topic)
   : log(topic, LEVEL_ERROR)
@@ -190,10 +174,6 @@ void wex::log::flush()
   {
     switch (m_level)
     {
-      case LEVEL_ERROR:
-        LOG(ERROR) << text;
-        break;
-
       case LEVEL_STATUS:
         // this is a wxMSW bug, crash in test -tc=wex::stc -sc=find
         if (text.find("%") == std::string::npos)
@@ -202,8 +182,24 @@ void wex::log::flush()
         }
         break;
 
-      default:
-        VLOG(m_level) << text;
+      case LEVEL_DEBUG:
+        BOOST_LOG_TRIVIAL(debug) << text;
+        break;
+      case LEVEL_ERROR:
+        BOOST_LOG_TRIVIAL(error) << text;
+        break;
+      case LEVEL_FATAL:
+        BOOST_LOG_TRIVIAL(fatal) << text;
+        break;
+      case LEVEL_INFO:
+        BOOST_LOG_TRIVIAL(info) << text;
+        break;
+      case LEVEL_TRACE:
+        BOOST_LOG_TRIVIAL(trace) << text;
+        break;
+      case LEVEL_WARNING:
+        BOOST_LOG_TRIVIAL(warning) << text;
+        break;
     }
   }
 }
@@ -221,65 +217,30 @@ wex::log wex::log::info(const std::string& topic)
   return log(topic, LEVEL_INFO);
 }
 
-void wex::log::init(int argc, char** argv)
+void wex::log::init(const std::string& default_logfile)
 {
-  // Load elp configuration from file.
-  const path elp(config::dir(), "conf.elp");
-
-  if (elp.file_exists())
+  if (m_initialized)
   {
-    el::Loggers::reconfigureAllLoggers(el::Configurations(elp.string()));
+    return;
   }
 
-  // We need to convert argc and argv, as elp expects = sign between values.
-  // The logging-flags are handled by syncped.
-  std::vector<std::string>                               v;
-  const std::vector<std::pair<std::string, std::string>> supported{
-    {"-m", "-vmodule"},
-    {"-D", "--default-log-file"},
-    {"-L", "--logging-flags"},
-    {"-V", "--v"},
-    {"--level", "--v"},
-    {"--logfile", "--default-log-file"},
-    {"--logflags", "--logging-flags"},
-    {"--x", "--v"}}; // for testing with verbosity: --v=9
+  logging::add_common_attributes();
 
-  for (int i = 0; i < argc; i++)
-  {
-    bool added = false;
+  logging::add_console_log(
+    std::cout,
+    logging::keywords::format = "%TimeStamp% [%Severity%] %Message%");
 
-    for (const auto& s : supported)
-    {
-      if (strcmp(argv[i], s.first.c_str()) == 0)
-      {
-        if (i + 1 < argc)
-        {
-          const std::string option(argv[i + 1]);
-          v.push_back(std::string(std::string(s.second) + "=" + option));
-          added = true;
-          i++;
-          break;
-        }
-      }
-    }
+  const path logfile(
+    config::dir(),
+    wxTheApp->GetAppName().ToStdString() + ".log");
 
-    if (!added)
-    {
-      v.push_back(argv[i]);
-    }
-  }
+  logging::add_file_log(
+    logging::keywords::file_name =
+      default_logfile.empty() ? logfile.string() : default_logfile,
+    logging::keywords::rotation_size = 10 * 1024 * 1024,
+    logging::keywords::format        = "%TimeStamp% [%Severity%] %Message%");
 
-  std::vector<char*> w;
-  for (const auto& arg : v)
-  {
-    w.push_back((char*)arg.data());
-  }
-  w.push_back(nullptr);
-
-  START_EASYLOGGINGPP(w.size() - 1, w.data());
-
-  info("verbosity") << level_string(el::Loggers::verboseLevel());
-  trace("log setup") << elp.string();
+  m_initialized = true;
 }
 
 wex::log wex::log::status(const std::string& topic)