+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 **v21.04**
 - added FindWEX.cmake to assist using wex library using cmake projects
+- boost::log lib instead of easylogging++ lib
+- c++20 standard used, c++17 available as branch
 
 **v20.10** *October 2, 2020*
 - clang-format support, and wex code follows these guidelines